backend/requirements.txt
Django>=5.2.7
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.9
boto3>=1.34.0
python-dotenv>=1.0.0
